+++ client/python/client.py
@@ -5,7 +5,7 @@ import json
 __author__ = "Tomasz Ducin"
 __email__ = "tomasz.ducin@gmail.com"
 __license__ = "MIT"
-__version__ = "1.1"
+__version__ = "1.2"
 
 class PyCachedClient(object):
     def __init__(self):