+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2 (2013-11-01)
+
+- http access (twisted instead of cherrypy)
+- javascript client added
+
 ## 1.1 (2013-10-26)
 
 - travis integration (nosetests for python, phpunit for php)